requirements.txt
torch>=1.8
torchvision
numpy
pandas
scikit-learn
matplotlib
Pillow
shap
lime
tqdm
opencv-python
